Refactor mbta code
