+++ src/mbta-tracker.py
@@ -10,122 +10,12 @@ from src.displays.adafruit import AdaFruit
 from src.data.fonts import default_font
 from src.displays.simulate import Simulate
 from typing import List, Tuple
+from src.programs.mbta import display_train_arrival_times
 
 try:
     import RPi.GPIO as GPIO
 except:
     print("Could not import RPi.GPIO, are you running in simulate mode?")
-
-# Example URLs
-# redline_centralsq_outbound_url = 'https://api-v3.mbta.com/predictions?filter[stop]=place-cntsq&filter[direction_id]=1&page[limit]=3'
-# redline_centralsq_inbound_url = 'https://api-v3.mbta.com/predictions?filter[stop]=place-cntsq&filter[direction_id]=0&page[limit]=3'
-try:
-    with open("credentials.txt", "r") as file:
-        api_key = file.read().strip()
-except:
-    api_key = None
-
-headers = {"Accept": "application/json", "x-api-key": api_key}
-
-
-def get_arrival_times(stop: str, direction: int, limit: int):
-    # Fetch
-    response = requests.get(
-        url=f"https://api-v3.mbta.com/predictions?filter[stop]={stop}&filter[direction_id]={direction}&page[limit]={limit}",
-        headers=headers,
-        auth=None,
-    )
-    # Stringify the promise to data
-    data = response.json()
-    print("x-ratelimit-remaining: ", response.headers["x-ratelimit-remaining"])
-    # We don't need to worry about 'null' data for the arrival_time because the station we're predicting is not a 'first stop' station
-    # If there is something wrong, we can use the 'schedule_relationship' field to figure out why.
-
-    # Get current time
-    currTime = datetime.datetime.now()
-    arrivalTimes = []
-    for stop in data["data"]:
-        # Convert fetched string to datetime format
-        time = datetime.datetime.strptime(
-            stop["attributes"]["arrival_time"], "%Y-%m-%dT%H:%M:%S-%f:00"
-        )
-        # seconds till arrival time
-        arrivalSecs = (time - currTime).total_seconds()
-        # minutes till arrival time
-        arrivalMins = round(arrivalSecs / 60)
-        if arrivalMins <= 0:
-            arrivalTimes.append("Arrived")
-        else:
-            arrivalTimes.append(str(arrivalMins) + "min")
-    return arrivalTimes
-
-
-"""
-*INFO*
--- Parameter 0 --
-stop: place-cntsq = "Central Square Station"
-stop: place-davis = "Davis Square Station"
-stop: place-portr = "Porter Square Station"
-
--- Parameter 1 -- 
-direction: 0 = Inbound
-direction: 1 = Outbound
-
--- Parameter 2 --
-limit: Number of next "x" arrival times you want to see.
-
-
-There are a list set of rules from the documentation that we should take into account (https://www.mbta.com/developers/v3-api/best-practices)
-1. If `status` is non-null:
-	Display this value as-is
-2. If `departure_time` is null:
-	Do not display this prediction, since riders won't be able to board the vehicle
-3. Calculate the number of seconds until the vehicle reaches the stop, by subtracting the current time from the arrival time (if available) or the departure time (if not); call this value "seconds"
-4. If seconds < 0
-	Do not display this prediction, since the vehicle has already left the stop
-5. If seconds <= 90, and the `status` of the associated `vehicle` is "STOPPED_AT", and the vehicle's `stop` is the same as the prediction's `stop`:
-	Display "Boarding" (abbrev. "BRD")
-6. If seconds is <= 30
-	Display "Arriving" (abbrev. "ARR")
-7. If seconds is <= 60
-	Display "Approaching" (abbrev. "1 min")
-8. Round the seconds value to the nearest whole number of minutes, rounding up if exactly in-between; call this value "minutes"
-9. If minutes > 20
-	Display “20+ minutes” (abbrev. “20+ min”)
-10. Display the number of minutes followed by "minutes" (abbrev. "min"). For example:
-	Up to 89 seconds: "1 minute" or "1 min"
-	90 to 149 seconds: "2 minutes" or "2 min"
-	150 to 209 seconds: "3 minutes" or "3 min"
-"""
-
-
-# TODO: Delete function
-def update_train_times():
-    while True:
-        print(
-            "Central Square - Red Line - Inbound - ",
-            getArrivalTimes("place-cntsq", 0, 3),
-        )
-        # print(
-        #     "Central Square - Red Line - Outbound - ",
-        #     getArrivalTimes("place-cntsq", 1, 3),
-        # )
-        print(
-            "------------------------------------------------------------------------------"
-        )
-
-
-def print_text(display, lines):
-    """Update the display with this, return immediatly"""
-
-    # lines: List[str] = ["Hello World,", "how are you?"]
-
-    pixels = copy.deepcopy(state.background)
-
-    pixels = draw_text(pixels=pixels, lines=lines)
-
-    display.display_matrix(pixels)
-
 
 def print_default_font(display):
     """Display the entire default font one page at a time,
@@ -322,8 +212,6 @@ if __name__ == "__main__":
         state.num_programs = 5
         state.num_modes = 10
 
-        display_inbound = True
-        begin_time = datetime.datetime.now()
         times = []
         loop_num = 0
         while True:
@@ -331,33 +219,8 @@ if __name__ == "__main__":
             start_time = time.time()
             
             if state.program == 0:
-                # lines = ["    Central SQ.", "Inbound", "10 min", "11 min"]
-                # print_text(display)
                 print('api-key : ', api_key)
-                curr_time = datetime.datetime.now()
-                if (curr_time - begin_time).total_seconds() > 10:
-                    begin_time = curr_time
-                    display_inbound = display_inbound ^ 1
-                if display_inbound:
-                    arrival_time_inbound = get_arrival_times("place-cntsq", 0, 2)
-                    lines_inbound = [
-                        "    Central SQ.",
-                        "Inbound",
-                        f"{arrival_time_inbound[0]}",
-                        f"{arrival_time_inbound[1]}",
-                    ]
-                    print_text(display, lines=lines_inbound)
-                else:
-                    arrival_time_outbound = get_arrival_times("place-cntsq", 1, 2)
-                    lines_outbound = [
-                        "    Central SQ.",
-                        "Outbound",
-                        f"{arrival_time_outbound[0]}",
-                        f"{arrival_time_outbound[1]}",
-                    ]
-                    print_text(display, lines=lines_outbound)
-                
-
+                display_train_arrival_times(display)
             elif state.program == 1:
                 display_image(display)
